Add files via upload
This code takes a video containing a time lapse of a biofilm with a radially propagating signal and converts the signal into polar coordinates and radial averages. The moments of the radial average signal are found, and used to calculate mean, std, kurtosis and skewness. The average radial signal is also fitted with Gaussians. (Requires wmean.m)